put structs in own files
