+++ src/value_type.rs
@@ -1,0 +1,8 @@
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub enum ValueType {
+  Str(String),
+  I32(i32),
+  I64(i64),
+  FloatingPoint(f64),
+  Boolean(bool),
+}
